Abstract helper functions - Handle state outside of each function - Hold user data at entry point to feed to rest of app
