--- src/components/UserDetails.tsx
@@ -1,1 +1,0 @@
-// Modal for user details display
