--- src/helpers/userQuotesHelper.ts
@@ -1,1 +1,0 @@
-// Logic for add, delete quote
