+++ src/models/Quote.ts
@@ -1,0 +1,6 @@
+export interface Quote {
+  quote: string;
+  author: string;
+  authorId: number;
+  id: number;
+}
