+++ src/models/User.ts
@@ -1,0 +1,4 @@
+export interface User {
+  id: number;
+  apiKey: string;
+}
